update codes, upgrade BLINKER NRF52 lib & readme
